requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
attrs==25.1.0
black==24.8.0
cachetools==5.5.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
embreex==2.17.7.post6
exceptiongroup==1.2.2
fonttools==4.55.8
fvcore==0.1.5.post20221221
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.7
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iopath==0.1.9
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lxml==5.3.0
manifold3d==3.0.1
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
pickle5==0.0.11
pillow==9.5.0
pip==24.2
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycocotools==2.0.7
pycollada==0.8
pyparsing==3.1.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
rtree==1.3.0
scipy==1.10.1
seaborn==0.13.2
setuptools==75.1.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
svg-path==6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tomli==2.2.1
tqdm==4.67.1
trimesh==4.6.1
typing_extensions==4.11.0
tzdata==2025.1
urllib3==2.2.3
werkzeug==3.0.6
wheel==0.44.0
xatlas==0.0.9
xxhash==3.5.0
yacs==0.1.8
